+++ changelog.md
@@ -5,6 +5,7 @@
 * Reimplemented battle stingy recovery as BaseCampaignEventListener. That should make it more resilient and also apply to fights originating from Domain probes
 * Disabled salvage stingy recovery in the tutorial to prevent soft-locks.
 * Changed stingy weapon recovery to support different values based on weapon tier.
+* Improved messaging illegality of items because of low standing
 
 # Version 4.0.0
 * Reimagined submarket changes. Now willingness to sell is based on multiple factors.