+++ changelog.md
@@ -1,11 +1,11 @@
-# Version 4.0.1
+# Version 4.1.0
 * Fixed contact level not providing standing
 * Fixed stringy weapon recovery mistakenly applying to players
 * Changed costMultiplierProduction to costMultiplierBuildings and corrected the description
 * Reimplemented battle stingy recovery as BaseCampaignEventListener. That should make it more resilient and also apply to fights originating from Domain probes
 * Disabled salvage stingy recovery in the tutorial to prevent soft-locks.
 * Changed stingy weapon recovery to support different values based on weapon tier.
-* Improved messaging illegality of items because of low standing
+* Improved messaging illegality of items caused by low standing
 
 # Version 4.0.0
 * Reimagined submarket changes. Now willingness to sell is based on multiple factors.